Moved test to mw folder
