+++ __tests__/app.test.ts
@@ -160,28 +160,4 @@ describe('Testing routes', () => {
         done()
       })
   })
-
-  it('should use the logger middleware', (done) => {
-    const originalConsoleLog = console.log;
-
-    console.log = (log) => {
-      expect(log.split(' ')[2]).toBe('GET');
-      console.log = originalConsoleLog;
-      done();
-    }
-    
-    const app = new App()
-    app.use(logger({ timestamp: { format: 'HH:mm:ss' } }))
-
-    const server = app.listen()
-
-    const request: any = supertest(server)
-
-    request
-      .get('/')
-      .expect(404)
-      .end(() => {
-        server.close()
-      })
-  })
 })
